#134 Move validateEnvironmentConfig to separate file.
