+++ web-frame/src/app-shared/config/interfaces/environment-config.interface.ts
@@ -16,15 +16,3 @@ export interface IEnvironmentConfig {
   clientSecret: string;
   devMode?: boolean;
 }
-
-export function validateEnvironmentConfig(config: IEnvironmentConfig): void | never {
-  if (!config.apiBaseUrl || config.apiBaseUrl === '') {
-    throw new Error('Invalid environment configuration, apiBaseUrl is missing.');
-  }
-  if (!config.appBaseUrl) {
-    throw new Error('Invalid environment configuration, appBaseUrl is missing.');
-  }
-  if (!config.appBaseUrl.endsWith('/')) {
-    throw new Error('Invalid environment configuration, appBaseUrl must end with a forward slash (/).');
-  }
-}
